Calc transition STAKING_RATIO at runtime, make conversion ratio 64bits
For mocknet and testing we now allow changing the staking requirement
at runtime (i.e. for testing different staking requirements and
conversion ratios) the staking ratio needs to be updated with the
appropriate values.

Allow a higher precision conversion ratio to allow for more fine grained
conversions. For the $OXEN to $SESH conversion we now defer to a 128 bit
multiply to avoid overflow for the conversion.

+++ src/cryptonote_core/sent_transition/detail.h
@@ -12,7 +12,7 @@ namespace oxen::sent::devnet {
 extern const std::unordered_map<std::string, eth::address> addresses;
 extern const std::unordered_map<crypto::public_key, crypto::ed25519_public_key> proper_ed_keys;
 extern const std::unordered_map<crypto::ed25519_public_key, eth::bls_public_key> bls_keys;
-extern const std::pair<std::uint8_t, std::uint8_t> conv_ratio;
+extern const std::pair<std::uint64_t, std::uint64_t> conv_ratio;
 extern const std::unordered_map<eth::address, std::uint64_t> transition_bonus;
 }  // namespace oxen::sent::devnet
 
@@ -20,7 +20,7 @@ namespace oxen::sent::testnet {
 extern const std::unordered_map<std::string, eth::address> addresses;
 extern const std::unordered_map<crypto::public_key, crypto::ed25519_public_key> proper_ed_keys;
 extern const std::unordered_map<crypto::ed25519_public_key, eth::bls_public_key> bls_keys;
-extern const std::pair<std::uint8_t, std::uint8_t> conv_ratio;
+extern const std::pair<std::uint64_t, std::uint64_t> conv_ratio;
 extern const std::unordered_map<eth::address, std::uint64_t> transition_bonus;
 }  // namespace oxen::sent::testnet
 
@@ -28,7 +28,7 @@ namespace oxen::sent::localdev {
 extern const std::unordered_map<std::string, eth::address> addresses;
 extern const std::unordered_map<crypto::public_key, crypto::ed25519_public_key> proper_ed_keys;
 extern const std::unordered_map<crypto::ed25519_public_key, eth::bls_public_key> bls_keys;
-extern const std::pair<std::uint8_t, std::uint8_t> conv_ratio;
+extern const std::pair<std::uint64_t, std::uint64_t> conv_ratio;
 extern const std::unordered_map<eth::address, std::uint64_t> transition_bonus;
 }  // namespace oxen::sent::localdev
 
@@ -36,7 +36,7 @@ namespace oxen::sent::mainnet {
 extern const std::unordered_map<std::string, eth::address> addresses;
 extern const std::unordered_map<crypto::public_key, crypto::ed25519_public_key> proper_ed_keys;
 extern const std::unordered_map<crypto::ed25519_public_key, eth::bls_public_key> bls_keys;
-extern const std::pair<std::uint8_t, std::uint8_t> conv_ratio;
+extern const std::pair<std::uint64_t, std::uint64_t> conv_ratio;
 extern const std::unordered_map<eth::address, std::uint64_t> transition_bonus;
 }  // namespace oxen::sent::mainnet
